v2.5 - 03/29/2009
Replacement of "float" type by "double" to match the relic version conventions.
Minor improvements.

+++ amsb.c
@@ -9,7 +9,7 @@
 int main(int argc,char** argv)
 {
 	char name[50];
-	float m0,m32,tanb,A0,sgnmu,mtop,mbot,alphas_mz,delta0;
+	double m0,m32,tanb,A0,sgnmu,mtop,mbot,alphas_mz,delta0;
 
 	if(argc<4) 
 	{ 
@@ -26,13 +26,13 @@ int main(int argc,char** argv)
   	} 
 	else  
   	{
-		sscanf(argv[1],"%f",&m0);
-     		sscanf(argv[2],"%f",&m32);
-     		sscanf(argv[3],"%f",&tanb);
-     		if(argc>4) sscanf(argv[4],"%f",&sgnmu); else sgnmu=1;
-     		if(argc>5) sscanf(argv[5],"%f",&mtop); else mtop=172.4;   
-     		if(argc>6) sscanf(argv[6],"%f",&mbot); else mbot=4.2;
-     		if(argc>7) sscanf(argv[7],"%f",&alphas_mz); else alphas_mz=0.1176;
+		sscanf(argv[1],"%lf",&m0);
+     		sscanf(argv[2],"%lf",&m32);
+     		sscanf(argv[3],"%lf",&tanb);
+     		if(argc>4) sscanf(argv[4],"%lf",&sgnmu); else sgnmu=1;
+     		if(argc>5) sscanf(argv[5],"%lf",&mtop); else mtop=172.4;   
+     		if(argc>6) sscanf(argv[6],"%lf",&mbot); else mbot=4.2;
+     		if(argc>7) sscanf(argv[7],"%lf",&alphas_mz); else alphas_mz=0.1176;
   	}	
 	sprintf(name,"amsb.lha");
