+++ nuhm.c
@@ -9,7 +9,7 @@
 int main(int argc,char** argv)
 {
 	char name[50];
-	float m0,m12,tanb,A0,mu,mA,mtop,mbot,alphas_mz,delta0;
+	double m0,m12,tanb,A0,mu,mA,mtop,mbot,alphas_mz,delta0;
 
 	if(argc<7) 
 	{ 
@@ -28,15 +28,15 @@ int main(int argc,char** argv)
 	} 
 	else  
 	{
-		sscanf(argv[1],"%f",&m0);
-		sscanf(argv[2],"%f",&m12);
-		sscanf(argv[3],"%f",&A0);
-		sscanf(argv[4],"%f",&tanb);
-		sscanf(argv[5],"%f",&mu);
-		sscanf(argv[6],"%f",&mA);
-		if(argc>7) sscanf(argv[7],"%f",&mtop); else mtop=172.4;   
-		if(argc>8) sscanf(argv[8],"%f",&mbot); else mbot=4.2;
-		if(argc>9) sscanf(argv[9],"%f",&alphas_mz); else alphas_mz=0.1176;
+		sscanf(argv[1],"%lf",&m0);
+		sscanf(argv[2],"%lf",&m12);
+		sscanf(argv[3],"%lf",&A0);
+		sscanf(argv[4],"%lf",&tanb);
+		sscanf(argv[5],"%lf",&mu);
+		sscanf(argv[6],"%lf",&mA);
+		if(argc>7) sscanf(argv[7],"%lf",&mtop); else mtop=172.4;   
+		if(argc>8) sscanf(argv[8],"%lf",&mbot); else mbot=4.2;
+		if(argc>9) sscanf(argv[9],"%lf",&alphas_mz); else alphas_mz=0.1176;
 	}	
 	sprintf(name,"nuhm.lha");
