+++ src/kmunu.c
@@ -1,21 +1,21 @@
 #include "include.h"
 
-float Kmunu_pimunu(struct parameters* param)
+double Kmunu_pimunu(struct parameters* param)
 /* computes the ratio BR(K-> mu nu)/BR(pi-> mu nu) */
 {
-	float Vus_Vud=0.2321;
-	float m_pi=0.1396;
-	float fK_fpi=1.189;
-	float delta_em=-0.0070;
-	float life_pi=2.6033e-8;
-	float life_K=1.2380e-8;  
+	double Vus_Vud=0.2321;
+	double m_pi=0.1396;
+	double fK_fpi=1.189;
+	double delta_em=-0.0070;
+	double life_pi=2.6033e-8;
+	double life_K=1.2380e-8;  
 
 #ifdef SMONLY
 	return param->m_K/m_pi*(1.+delta_em)*pow(Vus_Vud*fK_fpi*(1.-param->mass_mu*param->mass_mu/param->m_K/param->m_K)/(1.-param->mass_mu*param->mass_mu/m_pi/m_pi),2.)*life_K/life_pi;
 #endif
 	
-	float alphas_MSOFT=alphas_running(param->MSOFT_Q,param->mass_top_pole,param->mass_b_pole,param);
-	float epsilon0=-2./3.*alphas_MSOFT/pi*param->mu_Q/param->mass_gluino*
+	double alphas_MSOFT=alphas_running(param->MSOFT_Q,param->mass_top_pole,param->mass_b_pole,param);
+	double epsilon0=-2./3.*alphas_MSOFT/pi*param->mu_Q/param->mass_gluino*
 H2(param->MqL2_Q*param->MqL2_Q/param->mass_gluino/param->mass_gluino,param->MsR_Q*param->MsR_Q/param->mass_gluino/param->mass_gluino);
 	
 	return param->m_K/m_pi*(1.+delta_em)*pow(Vus_Vud*fK_fpi*(1.-param->mass_mu*param->mass_mu/param->m_K/param->m_K)/(1.-param->mass_mu*param->mass_mu/m_pi/m_pi)	*(1.-param->m_K*param->m_K/param->mass_H/param->mass_H*(1.-param->mass_d/param->mass_s)*param->tan_beta*param->tan_beta/(1.+epsilon0*param->tan_beta)),2.)*life_K/life_pi;
@@ -23,15 +23,15 @@ H2(param->MqL2_Q*param->MqL2_Q/param->mass_gluino/param->mass_gluino,param->MsR_
 
 /*--------------------------------------------------------------------*/
 
-float Rl23(struct parameters* param)
+double Rl23(struct parameters* param)
 /* computes the ratio Rl23 */
 {
 #ifdef SMONLY
 	return 1.;
 #endif
 
-	float alphas_MSOFT=alphas_running(param->MSOFT_Q,param->mass_top_pole,param->mass_b_pole,param);
-	float epsilon0=-2./3.*alphas_MSOFT/pi*param->mu_Q/param->mass_gluino*
+	double alphas_MSOFT=alphas_running(param->MSOFT_Q,param->mass_top_pole,param->mass_b_pole,param);
+	double epsilon0=-2./3.*alphas_MSOFT/pi*param->mu_Q/param->mass_gluino*
 H2(param->MqL2_Q*param->MqL2_Q/param->mass_gluino/param->mass_gluino,param->MsR_Q*param->MsR_Q/param->mass_gluino/param->mass_gluino);
 	
 	return fabs(1.-param->m_K*param->m_K/param->mass_H/param->mass_H*(1.-param->mass_d/param->mass_s)*param->tan_beta*param->tan_beta/(1.+epsilon0*param->tan_beta));
@@ -39,7 +39,7 @@ H2(param->MqL2_Q*param->MqL2_Q/param->mass_gluino/param->mass_gluino,param->MsR_
 
 /*--------------------------------------------------------------------*/
 
-float Kmunu_pimunu_calculator(char name[])
+double Kmunu_pimunu_calculator(char name[])
 /* "container" function scanning the SLHA file "name" and calculating BR(K-> mu nu)/BR(pi-> mu nu) */
 {
 	struct parameters param;
@@ -53,7 +53,7 @@ float Kmunu_pimunu_calculator(char name[])
 
 /*--------------------------------------------------------------------*/
 
-float Rl23_calculator(char name[])
+double Rl23_calculator(char name[])
 /* "container" function scanning the SLHA file "name" and calculating the ratio Rl23 */
 {
 	struct parameters param;
